// src/app/models/registros.ts
export class Registros{
  id_user?:string;
  id?: number;
  id_rg?: number;
  mot?: string;
  cnh_mot?: string;
  empresa_mot?: string;
  n_container?:number;
  destino_mot?: string;
  placa_cv?: string;
  placa_carreta?: string;
  lacres?: string;
  placa_carreta_saida?:string;
  lcr_saida?: string;
  id_inquilino?:number;
  data_entrada?: Date;
  hora_enviada?:Date;
  img_logo?: string;
  nome?:string;
  nome_aut_inquilino?:string;
  tipo_registro?: number;
  n_cont_saida?: string;
  tiporg?: number;
  noList?:string;

}